params: extract chain configs to separate files
